+++ packages/ui/src/index.ts
@@ -70,6 +70,11 @@ export {
 } from './lib/segmented-button/segmented-button';
 export { ElProgress, type ElProgressSize } from './lib/progress/progress';
 export { ElProgressCircle } from './lib/progress/progress-circle';
+export {
+  ElSlider,
+  type ElSliderSize,
+  type ElSliderThumb,
+} from './lib/slider/slider';
 export { ElAvatar, type ElAvatarSize } from './lib/avatar/avatar';
 export { ElCard, type ElCardAppearance, type ElCardSize } from './lib/card/card';
 export {